+++ src/main/java/com/caffeine/gwanghwamun/domain/menu/service/MenuService.java
@@ -11,7 +11,7 @@ import com.caffeine.gwanghwamun.domain.store.entity.Store;
 import com.caffeine.gwanghwamun.domain.store.repository.StoreRepository;
 import com.caffeine.gwanghwamun.domain.user.entity.User;
 import com.caffeine.gwanghwamun.domain.user.entity.UserRoleEnum;
-import com.caffeine.gwanghwamun.domain.user.security.UserDetailsImpl;
+import com.caffeine.gwanghwamun.common.security.model.UserDetailsImpl;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
