+++ src/main/java/com/caffeine/gwanghwamun/domain/user/controller/AuthTestController.java
@@ -1,6 +1,6 @@
 package com.caffeine.gwanghwamun.domain.user.controller;
 
-import com.caffeine.gwanghwamun.domain.user.security.UserDetailsImpl;
+import com.caffeine.gwanghwamun.common.security.model.UserDetailsImpl;
 import java.util.HashMap;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
